use fork of tinyexpr to make binary size smaller, use floats instead of doubles. add tinyexpr to project file so includes can be resolved. Don't download libs if already available

+++ BulbChipSTM32C071FBPx/setup.sh
@@ -7,7 +7,7 @@ TINYUSB_VERSION="0.18.0"
 TINYUSB_COMMIT="86ad6e56c1700e85f1c5678607a762cfe3aa2f47"
 
 # tinyexpr
-TINYEXPR_COMMIT="4a7456e"
+TINYEXPR_COMMIT="966948a116a7ccb0b4a632c01b909ad6d480d8bb"
 
 # Unity
 UNITY_VERSION="v2.6.1"
@@ -16,33 +16,45 @@ UNITY_COMMIT="cbcd08fa7de711053a3deec6339ee89cad5d2697"
 mkdir -p libs &&
 cd libs &&
 
-rm -rf lwjson &&
-echo "Downloading lwjson $LWJSON_VERSION ($LWJSON_COMMIT)..." &&
-curl -L "https://github.com/MaJerle/lwjson/archive/$LWJSON_COMMIT.tar.gz" > lwjson.tar.gz &&
-tar -xzf lwjson.tar.gz &&
-rm lwjson.tar.gz &&
-mv "lwjson-$LWJSON_COMMIT" lwjson &&
+if [ ! -d "lwjson" ]; then
+    echo "Downloading lwjson $LWJSON_VERSION ($LWJSON_COMMIT)..." &&
+    curl -L "https://github.com/MaJerle/lwjson/archive/$LWJSON_COMMIT.tar.gz" > lwjson.tar.gz &&
+    tar -xzf lwjson.tar.gz &&
+    rm lwjson.tar.gz &&
+    mv "lwjson-$LWJSON_COMMIT" lwjson
+else
+    echo "lwjson already exists, skipping download."
+fi &&
 
-rm -rf tinyusb &&
-echo "Downloading tinyusb $TINYUSB_VERSION ($TINYUSB_COMMIT)..." &&
-curl -L "https://github.com/hathach/tinyusb/archive/$TINYUSB_COMMIT.tar.gz" > tinyusb.tar.gz &&
-tar -xzf tinyusb.tar.gz &&
-rm tinyusb.tar.gz &&
-mv "tinyusb-$TINYUSB_COMMIT" tinyusb &&
+if [ ! -d "tinyusb" ]; then
+    echo "Downloading tinyusb $TINYUSB_VERSION ($TINYUSB_COMMIT)..." &&
+    curl -L "https://github.com/hathach/tinyusb/archive/$TINYUSB_COMMIT.tar.gz" > tinyusb.tar.gz &&
+    tar -xzf tinyusb.tar.gz &&
+    rm tinyusb.tar.gz &&
+    mv "tinyusb-$TINYUSB_COMMIT" tinyusb
+else
+    echo "tinyusb already exists, skipping download."
+fi &&
 
 python3 tinyusb/tools/get_deps.py stm32c0 &&
 
-rm -rf tinyexpr &&
-echo "Downloading tinyexpr ($TINYEXPR_COMMIT)..." &&
-curl -L "https://github.com/codeplea/tinyexpr/archive/$TINYEXPR_COMMIT.tar.gz" > tinyexpr.tar.gz &&
-tar -xzf tinyexpr.tar.gz &&
-rm tinyexpr.tar.gz &&
-mv "tinyexpr-$TINYEXPR_COMMIT"* tinyexpr &&
+if [ ! -d "tinyexpr" ]; then
+    echo "Downloading tinyexpr ($TINYEXPR_COMMIT)..." &&
+    curl -L "https://github.com/huntj88/tinyexpr/archive/$TINYEXPR_COMMIT.tar.gz" > tinyexpr.tar.gz &&
+    tar -xzf tinyexpr.tar.gz &&
+    rm tinyexpr.tar.gz &&
+    mv "tinyexpr-$TINYEXPR_COMMIT"* tinyexpr
+else
+    echo "tinyexpr already exists, skipping download."
+fi &&
 
-rm -rf Unity &&
-echo "Downloading Unity $UNITY_VERSION ($UNITY_COMMIT)..." &&
-curl -L "https://github.com/ThrowTheSwitch/Unity/archive/$UNITY_COMMIT.tar.gz" > Unity.tar.gz &&
-tar -xzf Unity.tar.gz &&
-rm Unity.tar.gz &&
-mv "Unity-$UNITY_COMMIT" Unity
+if [ ! -d "Unity" ]; then
+    echo "Downloading Unity $UNITY_VERSION ($UNITY_COMMIT)..." &&
+    curl -L "https://github.com/ThrowTheSwitch/Unity/archive/$UNITY_COMMIT.tar.gz" > Unity.tar.gz &&
+    tar -xzf Unity.tar.gz &&
+    rm Unity.tar.gz &&
+    mv "Unity-$UNITY_COMMIT" Unity
+else
+    echo "Unity already exists, skipping download."
+fi
